+++ src/main/java/com/example/harumeonglog/domain/auth/controller/AuthController.java
@@ -1,0 +1,33 @@
+package com.example.harumeonglog.domain.auth.controller;
+
+import com.example.harumeonglog.domain.auth.controller.specification.AuthControllerSpecification;
+import com.example.harumeonglog.domain.auth.dto.request.AuthRequest;
+import com.example.harumeonglog.domain.auth.dto.response.AuthResponse;
+import com.example.harumeonglog.domain.auth.service.AuthCommandService;
+import com.example.harumeonglog.domain.member.entity.Member;
+import com.example.harumeonglog.global.common.response.CustomResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/auth")
+public class AuthController implements AuthControllerSpecification {
+
+    private final AuthCommandService authCommandService;
+
+    @PostMapping("/{provider}/login")
+    public CustomResponse<AuthResponse.AuthLoginResponse> login(@PathVariable String provider, @RequestBody AuthRequest.AuthLoginRequest request) {
+        AuthResponse.AuthLoginResponse response = authCommandService.login(provider, request);
+        return CustomResponse.ok(response);
+    }
+
+
+    @PostMapping("/logout")
+    public CustomResponse<AuthResponse.AuthLogoutResponse> logout() {
+        // TODO: Annotation으로 변경 필요
+        Member member = Member.builder().build();
+        AuthResponse.AuthLogoutResponse response = authCommandService.logout(member);
+        return CustomResponse.ok(response);
+    }
+}
