+++ src/main/java/com/example/harumeonglog/domain/auth/controller/specification/AuthControllerSpecification.java
@@ -1,0 +1,19 @@
+package com.example.harumeonglog.domain.auth.controller.specification;
+
+import com.example.harumeonglog.domain.auth.dto.request.AuthRequest;
+import com.example.harumeonglog.domain.auth.dto.response.AuthResponse;
+import com.example.harumeonglog.global.common.response.CustomResponse;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+
+public interface AuthControllerSpecification {
+    @Operation(summary = "로그인 API by 서정모", description = "소셜 로그인")
+    @Parameter(name = "provider", description = "소셜 로그인 종류 (KAKAO, APPLE), APPLE은 현재 미완성 상태 요청 X")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "COMMON200", description = "성공입니다."),
+            @ApiResponse(responseCode = "AUTH400", description = "토큰이 유효하지 않습니다.")
+    })
+    CustomResponse<AuthResponse.AuthLoginResponse> login(String provider, AuthRequest.AuthLoginRequest request);
+}
