+++ src/main/java/com/example/harumeonglog/domain/auth/dto/response/AuthResponse.java
@@ -1,0 +1,21 @@
+package com.example.harumeonglog.domain.auth.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+public class AuthResponse {
+
+    @Getter
+    @Builder
+    public static class AuthLoginResponse {
+        private Long memberId;
+        private String accessToken;
+        private String refreshToken;
+    }
+
+    @Getter
+    @Builder
+    public static class AuthLogoutResponse {
+        private Long memberId;
+    }
+}
