+++ src/main/java/com/example/harumeonglog/domain/auth/service/OAuth2Service.java
@@ -1,4 +1,4 @@
-package com.example.harumeonglog.domain.member.service;
+package com.example.harumeonglog.domain.auth.service;
 
 import com.example.harumeonglog.global.security.domain.CustomUserDetails;
