+++ src/main/java/com/example/harumeonglog/domain/member/dto/response/MemberResponse.java
@@ -18,20 +18,6 @@ public class MemberResponse {
 
     @Getter
     @Builder
-    public static class MemberLoginResponse {
-        private Long memberId;
-        private String accessToken;
-        private String refreshToken;
-    }
-
-    @Getter
-    @Builder
-    public static class MemberLogoutResponse {
-        private Long memberId;
-    }
-
-    @Getter
-    @Builder
     public static class MemberInfoUpdateResponse {
         private Long memberId;
         private String image;
